Visual feedback for gun system
Added a second delegate & new event for gun system.
When fired, the gun raises the OnGunFired event, which will only be fired once per shot regardless of ammoPerShot value, and carries no data about the shot's direction, etc.
The OnGunShotFired event will be raised for each shot produced by the gun, with origin and direction information.
Fixed null reference exception in Arsenal script when no valid guns are equipped.
Added script PlayParticleOnGunFired which uses the new gun event to fire a particle along the trajectory of a gun shot.
Added template prefab for a basic gunshot event. This will probably need to be changed a lot to look nicer.