Tiny delay for gunshot particle emitter
Added a random tiny delay when emitting a particle for a gunshot, hopefully fixing the synchronicity issue.

Particles in Unity will only emit particles once per frame. Since the Emit function will emit in the forward direction of the transform, changing the direction and Emitting several times in a single frame would not make any difference, and the particles would emit in the direction of the last direction-change. Adding a random delay should almost always prevent this from happening.