Aiming off geometry fixes, camera max aim offset
Added a maximum offset distance to the CameraController.
Modified the behaviour of the PlayerAim script so that aiming persists even when the mouse is off geometry. When the mouse leaves geometry, the previous valid aim point is remembered, and a new point is found at the horizontal plane of the last. This behaviour may not be completely ideal and may have to be tweaked.