Damage data structure planning
Started implementing a new damage system. All damage will be done via a DamageData ScriptableObject instance rather than a simple float value. This will allow attackers to register to a "Damage Layer", and only be able to attack other particular layers.

Began implementation of DamageLayerManager which handles saving and loading the LayerNames file in the resources folder. 
Created a DamageLayerMask struct which can be used to specify which layers will be damaged by an attack.