Added Cursor static class, refactored aim code
Changed parameters for GameController.ITarget's ReceiveMoveInput & ReceiveAimInput functions. Removed the analog aim function. Control type is now tracked by a GameController, and passed to the target as a parameter in the aim function.

Added namespace StarForceRecon to NewPlayerController script.
Removed needlessly tracked variables from NewPlayerController & refactored code to run in a logical order. Aiming is now handled directly from the ReceiveAimInput function rather than being stored and handled in the next Update call.
Added Cursor static class under the StarForceRecon namespace, which stores a cursor position in Viewport space (0-1 both axes), and has a few simple functions to move the cursor around the screen.
AimPlayerCharacter now receives data from the input target functions and passes this through to the ModifyCursorPosition function. This function uses the new Cursor functions to move the cursor on the screen (with different behaviour for a mouse vs. joystick).