Joystick aiming
Last valid aiming-input time is now tracked for mouse & joystick aiming. 
When joystick has been used more recently than mouse to aim, aiming uses the 'joystick' aiming method. An aim destination is calculated around the viewport center relative to the direction of the analog stick. When the analog stick is released, the cursor will quickly reset back to the center of the viewport.
Added a function MoveTo to the Cursor class, which aims to smoothly move the cursor to the specified destination point.