Fixed InControlManager duplication bug
Adding the InControlManager component to a gameobject will add it to the object, but also create a second copy, due to the SingletonMonobehaviour class it inherits from. The GameController static constructor now uses the Instance function to create a single instance, rather than handling the containing object itself.
This fix in turn fixes the issue with WasPressed never returning true, fixing button press code.