Fixed damage-layer definition issue.
Fixed an issue where the damage-layer definition file would be recreated every time the editor refreshed. This was caused by using the Resources.Load method, which seems to only be useful at runtime. The issue was fixed by switching to the AssetDatabase.LoadAssetAtPath method.
Renamed the script while I was at it