Working masks & layers, definition creation
DamageData is no longer a ScriptableObject. It is now just a class with a custom propertyDrawer.
DamageLayerMask is now a contained class DamageLayer.Mask.
Added method GetLayerNames to Utils class to return an array of all non-blank layer names.
Added methods to get masks via params of int or string.
Added propertyDrawer for DamageData class.
Added propertyDrawer for DamageLayer, using a drop down which allows a single layer to be picked.
Added propertyDrawer for DamageLayer.Mask which uses a drop down MaskField to select a combination of layers.