DamageLayer.Modifier struct
DamageLayer now contains a Modifier struct which is used to set damage modifiers for particular layers. Stores layers in a dictionary, along with a float value.
Added overrides & overloaders for DamageLayer's Equals() function, GetHashCode() function, == operator and != operator so that it may be used as a key in Dictionaries properly. Also added implicit operators between int and DamageLayer.
DamageData now has a Modifiers struct. Getting the damage value now requires a layer parameter, which is used to apply the correct modifier before returning a damage value.
Started a PropertyDrawer for the new DamageLayer.Modifier struct. This should allow the modifier values for existing layers to be altered or removed, as well as any layers not currently contained to be added to the modifier struct.