Squad Follow, Prefab model loading
Added SquadFollow script for basic squad grouping. More work required to make this behave nicely.
Changed PlayerController to have a list of Behaviour rather than MonoBehaviour, for both AI and Player scripts. This allows things like NavMeshAgent to be toggled with character swapping.
Added InitializeChildResource script which aims to instantiate an object in Resources as a child of the transform. This is now used to add the mesh to the player character prefabs, which means only one prefab is needed for all characters.