Laser pointer fix
Added methods to AimHandler to get a copy of the layermask and valid tags array.
Added function to RaycastingHelper to get the first valid hit depending on a layermask and tags array.
LaserTest now casts a ray directly forward out of the origin, rather than relying on the aim script recording the last point.